fix: round-2 QA polish
Three of four polish items from the second QA pass; one skipped as a
deliberate convention choice.

Type safety:
- ContextAssembledSource and SkillsLoadedEntry are now defined once in
  src/engine/types.ts and re-exported from src/conversation/types.ts.
  ContextAssembledPayload.sources/excluded use the strict shape, so an
  emitter that ships a row without `tokens` (or any other required
  field) is a compile error rather than a runtime cast lie. Caught one
  loose typing in runtime.ts which has been tightened.

Dead-code:
- Removed the unreachable {workDir}/conversations fallback in
  getEventStore — the production path always returns a workspace-scoped
  event-sourced store, and tests mock through the same interface.

Authoring guide:
- Added a paragraph documenting which glob shapes the skills selector
  actually understands (`*`, `<prefix>__*`, `*__<suffix>`). Patterns
  with embedded wildcards (e.g. `*__patch_*`) fall back to literal
  exact-match and silently never load — surface this so authors don't
  ship dead skills.

Skipped: `errorResult` JSON-wrap vs plain text — the new plain-text
shape matches the AGENTS.md "human-readable summary in content"
convention better than instructions.ts's JSON-wrap; convergence should
go the other direction when instructions.ts is next touched.
