+++ src/engine/types.ts
@@ -163,24 +163,49 @@ export interface RunMetadata {
 }
 
 export interface SkillsLoadedPayload {
-  skills: Array<{
-    id: string;
-    layer: 3;
-    scope: "platform" | "workspace" | "user" | "bundle";
-    version: string;
-    tokens: number;
-    loadedBy: "always" | "tool_affinity";
-    reason: string;
-  }>;
+  skills: SkillsLoadedEntry[];
   totalTokens: number;
 }
 
 export interface ContextAssembledPayload {
-  sources: Array<Record<string, unknown>>;
-  excluded: Array<Record<string, unknown>>;
+  sources: ContextAssembledSource[];
+  excluded: ContextAssembledSource[];
   totalTokens: number;
   modelMaxContext?: number;
   headroomTokens?: number;
+}
+
+/**
+ * Per-skill telemetry attached to a `skills.loaded` event. Re-exported
+ * from `src/conversation/types.ts` so emitters and persisters reference
+ * one definition; drift surfaces as a type error.
+ */
+export interface SkillsLoadedEntry {
+  id: string;
+  layer: 3;
+  scope: "platform" | "workspace" | "user" | "bundle";
+  version: string;
+  tokens: number;
+  loadedBy: "always" | "tool_affinity";
+  reason: string;
+}
+
+/**
+ * One entry in `context.assembled.sources` / `excluded`. Required `tokens`
+ * + free-form discriminators (`count`, `version`, `turns`, etc.) per source
+ * kind. Tightening the engine payload to this shape (vs `Record<string,
+ * unknown>`) prevents emitters from accidentally shipping rows without a
+ * token count.
+ */
+export interface ContextAssembledSource {
+  kind: string;
+  count?: number;
+  tokens: number;
+  toolSetHash?: string;
+  version?: string | number;
+  userId?: string;
+  turns?: number;
+  compacted?: boolean;
 }
 
 /**
